Introducing tabs.windowSize()
This lets a script set the window size (width, height). This a easy to implement alternative for the mobile emulation that we had in v1 (the api for this only present in XUL, not in WebExt). However the downside is that we are setting the window size instead of the viewport size. And we can not set all sizes because of OS limitations (minimum window size, actual screen resolution, etc)