Fix for "Could not establish connection. Receiving end does not exist." error occuring during script runs
The error is thrown here: mozilla-central/toolkit/components/extensions/ExtensionChild.jsm : 502

```javascript
    this._sendMessage(messageManager, "Extension:Connect", msg, recipient).catch(error => {
      if (error.result === MessageChannel.RESULT_NO_HANDLER) {
        error = {message: "Could not establish connection. Receiving end does not exist."};
      } else if (error.result === MessageChannel.RESULT_DISCONNECTED) {
        error = null;
      }
      port.disconnectByOtherEnd(new StructuredCloneHolder(error));
    });
```

The case where a content frame is in the process of unloading was already handled. This error should be handled in the same way, it occurs when a new frame is being loaded, but our listener has not been attached yet. When `tab.wait()` encounters `RPCNoResponse` it will repeat the script when the content script notifies us that it is ready.