Fix for integration tests being very slow and periodically failing in recen firefox versions
I suspect that firefox changed something with their Garbage Collecting and we are navigation to pages so fast that firefox has no time to clean up its garbage. In this fix we simply restart the browser every 5 runs. An alternative would be to wait a few seconds after every test, however this would slow down the whole run very much. In the old extension API it was actually possible to trigger a GC cleanup manually, and this is what the old performr-runner used, however this is no longer possible.