Move the script-env Worker to its own special tab in the script window
Previously this Worker was created by the background script. But now whenever a script is being run, a special tab (pointing to a blank extension html page) is opened in the script window.
The benefit of this is the script now runs in the same contextual identity as the other tabs, but still without any cross origin checks because it is an extension page. Meaning, scripts can use the global fetch() function to perform api calls without origin checks, but also without polluting the default cookie store which would persist between runs.