Rework build scripts, bootstrapping of the extension and launching firefox
There is now a simple javascript class to launch Openrunner with, when used as a node.js library. It takes care of building the extension and profile, temporary directories, starting firefox, and running any amount of scripts.

There is now an option to load the extension using the debugging protocol instead of preloading it into the profile. This can be used to start Openrunner using a regular firefox release, because it bypasses any extension signing restrictions.

There is also a new build command which gives a ready made extension as a folder or as an .xpi file. And the other build commands have been changed so that they can be used standalone, instead of requiring the output of a previous build command.